decomp: two more ov022 parks, same 64-bit reading; a third improved
Gate green: DSD_OK=306.  Both matched functions already carried Ghidra names from an earlier
session (MergePendingBits_020ad474, SetOrClearNodeFlags_020ad838), so those were kept.

func_ov022_020ad474 (116 B) and func_ov022_020ad838 (168 B), both reloc-free, both now in
auto/.  Between them their parks blamed a "dead `& 0` flag term", a no-op self-store,
redundant stores that -O4 would remove, and a build-time 0-valued macro in a header we do not
have.  All of it was paired words being one 64-bit field.  ad474 matched on the first
compile once written that way.

ad838 took a second: 0x2000 sits in the HIGH half, so the constant is `0x2000LL << 32` =
0x200000000000.  Written unshifted it gives `orr r2, r1, #2` where the ROM has
`orr r2, r1, #0x2000` -- 4 bytes.  Which half carries the mask is also what tells you where
the field starts, and that is now its own line in the catalogue entry.

func_ov022_020ad4e8 improved but not closed: the guard on +0x464 is a 64-bit AND, and with
that it is the right SIZE (160) instead of 20 bytes short.  What is left is the
predicate-versus-branch residue its note already described, and an explicit if/else in either
direction does not help (27 lines one way, 4 bytes over the other).  The C in the park now
carries the fix so the next attempt starts from the better position.

Running total on this signature: seven functions across five parks, plus the twelve from the
three families this morning.  Four separate park notes had reached for a missing header or an
unreproducible construct, and every one of them was reading half of a 64-bit operation.

+++ src/overlays/ov022/nonmatching/func_ov022_020ad4e8.c
@@ -1,6 +1,16 @@
 /*
  * [nonmatching: predicate-vs-branch -- 140B vs 160B]
  *
+ * IMPROVED 2026-07-20: the guard on +0x464 is a 64-BIT AND, not a 32-bit one.  With
+ * `*(long long *)(arg0 + 0x464) & 0x20000000` the function is the right SIZE (160) instead
+ * of 20 bytes short, and the residue drops to the register/predication difference the note
+ * below describes.  The C now carries that fix.
+ *
+ * Still open, and the note's original diagnosis stands for it: mwcc predicates the abs-delta
+ * block (rsbhi/lslhi/lsrhi) and pushes two extra registers where the ROM branches it out of
+ * line with a `movle` on the fall-through.  An explicit if/else in either direction is not
+ * better (27 lines one way, 4 bytes over the other).
+ *
  * Clamps a wrapped 16-bit angle delta toward a limit. mwcc predicates the
  * abs-delta block (rsbhi/lslhi/lsrhi) and the multi-&& guards inline where the
  * original branches them out of line (20B shorter). Equivalent; left uncarved.
@@ -18,7 +28,7 @@ unsigned int func_ov022_020ad4e8(int arg0, unsigned int arg1) {
     lim = *(int *)(arg0 + 0x4b0);
     if ((int)ad <= lim) return arg1;
     if ((*(unsigned int *)(arg0 + 0x24) & 4) != 0 &&
-        (*(unsigned int *)(arg0 + 0x464) & 0x20000000) == 0 && 0x7000 < ad) {
+        (*(long long *)(arg0 + 0x464) & 0x20000000) == 0 && 0x7000 < ad) {
         return arg1;
     }
     if (0x8000 < d) lim = -lim;
